pdfProcess/requirements.txt
# Python PDF Splitting Worker Requirements
# Core dependencies for the PDF splitting worker

# Async HTTP client
aiohttp>=3.8.0

# Async file operations
aiofiles>=23.0.0

# RabbitMQ client
pika>=1.3.0

# PDF processing
PyPDF2>=3.0.0

# OSS/S3 support (optional, for real storage)
oss2>=2.17.0
boto3>=1.26.0

# UUID generation
uuid>=1.30

# Logging and configuration
python-dotenv>=1.0.0

# Elasticsearch support for logging
elasticsearch>=8.0.0
requests>=2.28.0
urllib3>=1.26.0

# Development and testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0